Fully Working Real estate page
Leases and properties connected to backend, and delete and create properly. Changed monthly income of property to come from sum of leases rather than input on creation. Changed total income to come from sum of that. Maybe one day we can fix monthlyincome in the database to be calculated, but prolly not. Leases are linked to Properties by having a property_id field in the database matching an id of a property in properties table.